requirements.txt
streamlit
opencv-python-headless==4.9.0.80
numpy<2
scikit-image
scikit-learn
Pillow